Classification of diseased leaves using MatLab
Currently we are able to classify the leaves into 3 clusters in which each cluster contains background, diseased part and part with no disease. First, the contrast of the image is enhanced for better processing of image, then the image is segmented using otsu method. For using otsu method, the rgb image has to be converted to hsi colour space. rgb2hsi function is used in that purpose. For extracting features from the leaf image, svm classifier is used which could create 3 clusters each containing one of the 3 parts mentioned above. But this could not be used in the project anymore as it is very difficult to automate this process using MatLab.